feat(booking)add payment

+++ src/main/java/com/group3/evproject/entity/Booking.java
@@ -5,6 +5,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -50,7 +51,7 @@ public class Booking extends BaseEntity {
     LocalDateTime startTime;
 
     @Column(name = "reservation_fee")
-    Double reservationFee;
+    BigDecimal reservationFee;
 
     @Column(name = "end_time")
     LocalDateTime endTime;
